backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
itsdangerous==2.1.2

# JWT Token validation for PKCE flow
python-jose[cryptography]==3.3.0
httpx==0.27.0

# Azure SDKs
azure-identity==1.15.0
azure-mgmt-resource==23.0.1
azure-mgmt-authorization==4.0.0
azure-mgmt-costmanagement==4.0.1
azure-storage-blob==12.19.0
msgraph-sdk==1.2.0

# Pydantic for settings and validation
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Utilities
python-dotenv==1.0.0
